Add Plugin for PF Firewall
I've made a pair of plugins for DenyHosts that add and remove the
offending IP address to/from a table in a PF firewall (the firewall
for OpenBSD, FreeBSD and NetBSD). It also adds and removes them
to/from a persist file for the table. The plugins are written in
Python (they are my first Python scripts, so probably could be
improved).